Added code to handle OAuth and create jwt tokens

+++ src/controllers/oAuthController.ts
@@ -3,7 +3,9 @@ import prisma from "../db/db";
 import config from "../config/config";
 import { NotFoundError } from "../utils/exceptions/notFoundError";
 import { ClientError } from "../utils/exceptions/clientError";
-import { getGoogleTokens } from "../middlewares/oAuthMiddleware";
+import { getGoogleTokens, getGoogleUserInfo } from "../middlewares/oAuthMiddleware";
+import { CustomError } from "../utils/exceptions/customError";
+import { issueTokens } from "../middlewares/authMiddleware";
 
 class OAuthController {
 
@@ -55,20 +57,90 @@ class OAuthController {
         // Get access and refresh token from google
         const response = await getGoogleTokens(code as string);
         const data = response.data;
-        const accessToken = data.access_token;
-        const refreshToken = data.refresh_token;
-        const expiresIn = data.expres_in;
+        const googleAccessToken = data.access_token;
+        const googleRefreshToken = data.refresh_token;
+        const googleExpiresIn = data.expres_in;
 
         // check if old refresh token exists in cookie
         const oldRefreshToken = req.cookies?.refresh_token;
-        if (refreshToken) {
-            await prisma.refreshToken.delete({ where: { token: oldRefreshToken } })
+        if (oldRefreshToken) {
+            const oldTokenResult = await prisma.refreshToken.findUnique({ where: { token: oldRefreshToken } });
+            if (oldTokenResult) await prisma.refreshToken.delete({ where: { token: oldRefreshToken } });
             res.clearCookie('refresh_token', { httpOnly: true, sameSite: 'strict', secure: true });
         }
 
-        // 
+        // get the userinfo from google
+        const userInfo = await getGoogleUserInfo(googleAccessToken);
+        if (!userInfo) throw new NotFoundError("Cannot get user info. Login again.");
 
+        /*
+        {
+            "id": "112345678901234567890",
+            "email": "john.doe@gmail.com",
+            "verified_email": true,
+            "name": "John Doe",
+            "given_name": "John",
+            "family_name": "Doe",
+            "picture": "https://lh3.googleusercontent.com/a-/AOh14Gh9...",
+            "locale": "en"
+        }
+        */
+        // check if user exists then get info
+        let user = await prisma.user.findUnique({ where: { email: userInfo.email as string } });
+
+        // if user does not exists create new
+        if (!user) {
+            // create user
+            user = await prisma.user.create({
+                data: {
+                    email: userInfo.email,
+                    name: userInfo.name,
+                    avatarUrl: userInfo.picture,
+                }
+            })
+
+            if (!user) throw new CustomError("Error creating user. Try again.");
+
+            // create oauthaccounts
+            const resultOAuthAccounts = await prisma.oAuthAccount.create({
+                data: {
+                    provider: "google",
+                    providerAccountId: userInfo.id,
+                    userId: user.id,
+                    expiresAt: new Date() + googleExpiresIn,
+                }
+            })
+
+            if (!resultOAuthAccounts) {
+                await prisma.user.delete({ where: { email: userInfo.email } });
+                throw new CustomError("Error creating OAuth Account for user. Try again.");
+            }
+        }
+
+        // issue jwt access and refresh token
+        const { accessToken, newRefreshToken } = issueTokens(user.id, user.email);
+        if (!accessToken || !newRefreshToken) throw new CustomError("Unable to create access tokens.");
+
+        // store refresh token in db
+        const refereshTokenResult = await prisma.refreshToken.create({
+            data: {
+                token: newRefreshToken,
+                expiresAt: new Date(Date.now() + 30 * 24 * 60 * 60 * 1000),
+                userId: user.id,
+            }
+        });
+
+        // store refresh token in http only token
+        res.cookie(config.refreshToken as string, newRefreshToken, {
+            httpOnly: true,
+            secure: true,
+            sameSite: 'strict',
+            maxAge: 30 * 24 * 60 * 60 * 1000,
+        });
+
+        // send response containing the access token
+        res.status(200).json({ accessToken: accessToken });
     }
 }
 
-export default OAuthController; 87
+export default OAuthController;
